Query all cloudtrail events for stop and terminate
This commit uses the Paginator for CloudTrail events, instead of just
running a single query.

This should prevent missing events, even though it might mean longer
runtimes for CAD in case of very many events matching stop/terminate
events.

It also will only retry the event-gathering instead of the all the
actions that follow on the event, to keep the retry-loop smaller.