We tend to see a lot of clusters that have CAD run against them hit timeouts with egress.
After reviewing, I noticed we had set the timeout to 2 seconds which felt pretty tight.

I have updated that value to be 5 seconds, which lines up with what OCM uses to perform pre-cluster
creation network verification. I think unifying on a default (in osd-network-verifier next) will
ensure we are consistent about this across tools and services.