Isolate metrics and logging for investgate command
As a preparation step for adding new commands to cadctl, this commit
moves the logging and metrics code from the root command to the
investigate command. This allows for a cleaner separation of concerns
and makes it easier to add new commands in the future.

The initial logging only bare minimum, and expands
based on the command.

Signed-off-by: Petr Kotas <pkotas@redhat.com>