Use a builder pattern to initialize resources used by investigations.
To allow not all resources to be prefilled as not every investigation
needs everything, this introduces a Builder pattern that is prefilled
with a subset of fields (OCM, PD, Cluster, Logging) that can be extended
by each investigation.

Once everything that is supposed to be build is set, calling Build()
will return the requested resources.